// store/interfaces/user/userInfo.ts
interface userInfo {
    name: string;
    family: string;
    mobile: string;
    phone: string;
    email: string;
    type: string;
    token: string;
    password: string;
    address: string;
    sheba: string;
    bank: string;
    zipcode: string;
    state: string;
    city: string;
    gender: string;
    nationalCode: string;
    idNumber: string;
    commercialCode: string;
  }
  export default userInfo;
  